Fix psid cookie older client failure bug [PG-93]
Older clients, 2.5.x and below, send an ACK_V1 packet in response to
the server's HARD_RESET packet whereas 2.6.x clients send a CONTROL_V1
packet.  The code that checked the packet length of the client's
response failed to comprehend the fact the ACK_V1 packet does not
include a packet id field following the peer session id field.  So the
code rejected the ACK_V1 packet as being too short.

The fix was to require packet length only up to and including the peer
session id field.  This works to allow safe parsing for both the
ACK_V1 response and the CONTROL_V1 response.

Signed-off-by: Mark Deric <jmark@openvpn.net>