Untangle log header macro/class globs
The purpose of this change is to allow headers that require the
logbase.hpp classes to compile in executables using logsimple.hpp.

By munging classes and macros into both headers, an avoidable conflict
of macro re-definition is created.  This commit separates the classes
from the macros into new headers. Then propagates the mistake into the
current headers so none of the existing code is broken. ;-)

Signed-off-by: Mark Deric <jmark@openvpn.net>