Introduce more flexible logging class
This template base class allows specifying a maximum log level that
can be used to make the logging function empty above that level.

This allows us to make our logging more structured and move to a more
standard logging approach. The use of if constexpr ensures that logging of
very high level is not emitting code.

This also add a few macros (LOG_VERBOSE, LOG_INFO, LOG_TRACE) that are drop
in replacements for the old OPENVPN_xx_LOG and OPENVPN_xx_LOG_VERBOSE macros
and also ensure not to do string construction if logging at a certain level
is already disabled at compile time.

Signed-off-by: Arne Schwabe <arne@openvpn.net>