dco: Fix incorrect #ifdef vs #if usage for ENABLE_KOVPN
The code in dco/dcocli.hpp used #if ENABLE_KOVPN, which will
fail on newer compilers if the macro is defined in a source
file.  Compilers may not complain if the macro is defined on
the command line, via -D.

This type of checks should use either #ifdef or #if defined(...).
The #if conditional expects a boolean expression.

Since these code blocks also depended on #elif (also expects
a boolean expression , the defined(...) approach was chosen
throughout this file.

Signed-off-by: David Sommerseth <davids@openvpn.net>