Fix client logging problems
A refactoring of the logging code in commit 9ffa263b removed
conventional #ifndef header guards and replaced them with #pragma
once.  Some consumers of the code used the header #defines to guide
their behavior.  The missing guards caused the consumers to log
incorrectly.  The fix was to still rely on the #pragma once for
guarding, but to reinstate the #defines from the original headers.

Signed-off-by: Mark Deric <jmark@openvpn.net>