jsonhelper: Extend disabling -Wdangling-reference to GCC 14+
According to the notes in the issues and changelog GCC
has fixed some false-positives of this but basically
accepted the rest and will not fix them.

They introduced gnu::no_dangling attribute to allow people
to override this warning on a class/function basis. But
that doesn't help us since we also need to support older
compilers.

Signed-off-by: Frank Lichtenheld <frank@lichtenheld.com>