WFP: block DNS traffic to loopback conditionally
Before this commit traffic to loopback was limited when only DNS
(port 53) was blocked, due to the "not loopback" match condition being
replaced instead of the match condition being made more specific.

This broke the client option to override access to DNS servers listening
on loopback.

To fix this three things are done:
  1) do not add DNS block rules if the override option is active.
  2) explicitly block port 53 on loopback, except when the override
     option is active.
  3) remove the implicit block of port 53 on loopback and instead let
     the firewall rule for non-loopback devices only.

Signed-off-by: Heiko Hund <heiko@openvpn.net>