Change license of Core library to MPL-2.0 + AGPLv3
This change allows using the OpenVPN core library as library in
other projects. More specifically it allows others to use this
library in App Stores like the Apple App store.

Refer to LICENSE.rst and COPYRIGHT.MPL2 for details about the new
license.

This commit does not add license header to files that currently do
not have one (cmake build files, some of the unit test program
and generally non C/C++ files).

For the xkey files that are shared with OpenVPN 2.x the license also
allows using the GPL-2.0-only WITH openvpn-openssl-exception.

This keeps the AGPLv3 license as an option for existing projects
that want to stick to the AGPLv3 license instead of a more liberal
(and AGPLv3 compatible) MPL 2.0.

This also changes the Copyright lines in the affected files for OpenVPN
to no longer have an end date but rather just state (C) startyear-
without end year to indicate an ongoing copyright.

Signed-off-by: Arne Schwabe <arne@openvpn.net>