cliopt: Pass allowLocalLanAccess to TunProp::Config
The allowLocalLanAccess config setting was overlooked in the refactoring
happening in commit 0c5e356075cb17f3b1 which arrived in OpenVPN 3 Core
Library v3.9.

This results in the TunProp::Config::allow_local_lan_access being set to
the default value, regardless of what the
ClientAPI::ConfigCommon::allowLocalLanAccess setting was set to.  This
caused a regression to users depending on this feature, since the VPN
client implementation can only access this setting via the
ClientAPI::ConfigCommon object.

Signed-off-by: David Sommerseth <davids@openvpn.net>