mac/gw.hpp: refactor MacGatewayInfo implementation
In order simplify and to make it work for "add bypass route" functionality:

 - remove RTA_NETMASK flag. This is only needed to get the netmask
that applies to the route. For the purposes of getting the next-hop
gateway and netmask of the specific interface it is not needed. Moreover,
with this flag set by some reasons I got the default gateway instead of
specific host route.

 - remove OPENVPN_NEXTADDR(6) macros. Plain assignment and explicit length
adjustment is is cleaner.

 - remove memset, since brace initialization is enough.

 - define variables close to usage instead of at the top.

Signed-off-by: Lev Stipakov <lev@openvpn.net>