win: apply DNS settings for dhcp-option as well
Create an Action to apply DNS settings. Previously applying them was
implicitly done in the DNS::Action classes, which led to them not being
applied on the code path for handling DNS related dhcp-options. This
turned out to be an issue for some edge case with group policy related
NRPT rules.

Now applying them is done explicitly and is also more visible.

Jira: OVPN3-1272
Signed-off-by: Heiko Hund <heiko@openvpn.net>