mac/client/tunsetup.hpp: don't add bypass routes when redirect-gw and remote IP versions mismatch
Commit

 053600556 ("mac/client/tunsetup.hpp: switch to MacGatewayInfo implementation")

ensured that we use the proper gateway to reach remote. However, this
backfires in case when IP versions mismatch between remote and
redirect-gw. For example, when doing redirect-ipv6 we attempt to
get a best gateway to reach a remote and assume that remote is also
IPv6.

Fix by adding bypass route only when redirect-gw and remote IP protocols
match.

While on it, remove seem to be useless commented out code.

OVPN3-1274

Signed-off-by: Lev Stipakov <lev@openvpn.net>