win: add NRPT GPOL notifications for x86 Windows
The signature for RtlPublishWnfStateData() differs on 32-bit systems.
Here the state name needs to be given in two 32 bit values instead of
one 64 bit value. It also uses __stdcall as calling convention.

Find out if we're running in an 32 bit environment and call the
notification function in the expected way.

Jira: OVPN3-1280
Signed-off-by: Heiko Hund <heiko@openvpn.net>