mac: make routing socket write error non-fatal
Commit

  0536005 ("switch to MacGatewayInfo implementation")

switched to a new routing socket implementation for
determining the best gateway. When handling exclude routes,
we attempt to retrieve the default gateway. However,
the implementation currently throws a fatal error if
there is no default gateway, which can occur when
there is no IPv6 connectivity or, for example,
no default gateway in IPv4.

This lack of a default gateway leads to a confusing
routing socket write error. This commit fixes the issue
by making this error non-fatal and simply logging
an error message. As before, exclude routes
are not processed if there is no default gateway.

OVPN3-1281
GitHub: https://github.com/OpenVPN/openvpn3/issues/344

Signed-off-by: Lev Stipakov <lev@openvpn.net>