buffer: Add realign() member function
This function is used to realign the buffer so that the data starts at
a given offest from the beginning of the buffer. This is useful when
a struct is streamed into the buffer but it is not aligned to the
proper memory boundary.

Might also be useful for other use cases.

ConstBufferType::realign() is a const member function that does a
similar thing but since ConstBufferType doesn't own its buffer
it can't grow if this is required. This overload of realign() will
grow the buffer if that is needed, and if not it calls the const
version of realign().

Unit tests included.

Signed-off-by: Charlie Vigue <charlie.vigue@openvpn.com>