ProtoUnitTest: add unit test for TLS version mismatches
Add a simple unit test making sure that communication stops when
there's a TLS version mismatch between client and server.

Unfortunately, this does not save us from having to worry about
OpenVPN 3.x <-> OpenVPN 2.x connections, and it doesn't even
tell us if the server or the client raised an exception (since
they're both in the same process, the first one to do so will
break the connection attempt).

Having said that, it is probably still better to have a unit test
for this than not.

Signed-off-by: Razvan Cojocaru <razvan.cojocaru@openvpn.com>