httpserv: Reinstate move for buffer in httpserv
Coverity rightfully complains that the move operation in add_to_pipeline
might result in UB since it's performing a move on a reference to a
buffer from a potentially distant part of the pipeline. This change
simply ensures that the BufferAllocated type will not invoke
undefined behavior when used after a move.

Unit tests included to ensure this behavior says as is.

NOTE: The original code was correct, but the Coverity warning was
valid. This change is to ensure that the code is correct and that
we won't inadvertently introduce UB in the future. Next we can note
this PR and the ticket in the Coverity issue and ignore it safely.

Signed-off-by: Charlie Vigue <charlie.vigue@openvpn.com>