Buffer: Annotate test_buffer.cpp for coverity
This change adds coverity specific tags before each unit test line that
intentionally accesses the buffer after it was moved from. This is to
suppress coverity warnings about the buffer being accessed after it was
moved from. This is done because ensuring use-after-move works correctly
is a part of the test.

These comments should inform both Coverity and maintainers of this.

Signed-off-by: Charlie Vigue <charlie.vigue@openvpn.com>