Fix openvpn::is_openvpn_protocol()
The original version just assumes that the packet length will
always be less than 255 (i.e. that the most significant byte of len
will be 0), which is not always the case for tls-crypt-v2.

This patch uses the OpenVPN 2.x logic and especially the comments,
which are very useful to make sense of an otherwise more opaque
algorithm.

While at it, removed the openvpn/common/size.hpp dependency by
using cstddef / std::size_t.

Signed-off-by: Razvan Cojocaru <razvan.cojocaru@openvpn.com>