Make prepend_alloc try harder
Previously prepend_alloc would only try to allocate headspace if the
empty space at the beginning of the buffer was available. This change
maintains this as the preferred behavior but also allows for the
possibility of moving existing data down the buffer if this will make
the required space available.

This change is motivated by the fact that the current behavior can lead
to an exception being thrown if there are plenty of bytes available at
the end of the buffer but not at the beginning. This can happen if
.realign(0) is called on a buffer that will later have data prepended to
it.

Signed-off-by: Charlie Vigue <charlie.vigue@openvpn.com>