tests: Fix test_proto.cpp build for USE_TLS_AUTH and USE_TLS_CRYPT
A series of refactorings has left test_proto.cpp broken when built
with USE_TLS_AUTH and USE_TLS_CRYPT.

This change fixes the build, but the USE_TLS_CRYPT build will fail
when launched - this is probably because it attemtps to use the
TLS-auth key. A follow-up patch fixing that as well will need to
happen.

Signed-off-by: Razvan Cojocaru <razvan.cojocaru@openvpn.com>