Fix route removal issue on macOS when pushed route exists
macOS requires routes to have a unique destination/netmask combination.
As a result, if the server pushes a route that already exists,
it will be ignored. However, on disconnect, the existing route
with the same destination and netmask will be removed,
potentially affecting local connectivity.

To prevent this issue, we apply the destination/netmask combination
as a mark to both the add and delete route commands.
If adding a route fails, we remove the corresponding "delete route"
command from the undo list.

OVPN3-1283

Signed-off-by: Lev Stipakov <lev@openvpn.net>