Use default flag value many places
In this commit we start using the default flag value in many places
where we wanted a value of NO_FLAGS, thus if we want no flags, we
simply pass no flags argument. This is a follow up to the previous
commit where we added the default flag value to the Buffer class and
made it strongly typed.

Signed-off-by: Charlie Vigue <charlie.vigue@openvpn.com>