Unit tests: Take fuzzing-caused failures into account
Use test_retry() instead of just test() for ProtoUnitTest TLS
crypt V2 testing, like most other tests also do. This is required
because of the built-in fuzzing in test_proto.cpp, where we expect
some failures but consider that a test is passing if it eventually
runs with no errors.

Signed-off-by: Razvan Cojocaru <razvan.cojocaru@openvpn.com>