Add PSID/HMAC cookie support for TLS crypt V2
For TLS crypt V2 we have a CONTROL_HARD_RESET_CLIENT_V3 first
packet (with an appended WKc, holding potential server key
information), to which we reply with a synthetic (manually
built) CONTROL_HARD_RESET_SERVER_V2 holding our generated
session ID, to which we expect a CONTROL_WKC_V1 packet, which
also contains a WKc that - assuming its session ID meets our
challenge - we must then use to set up the symmetric key for
the rest of the session.

For all this machinery to work, unwrap_tls_crypt_wkc() needed
to be updated in order for it to be able to handle CONTROL_WKC_V1,
as well as extracting the symmetric key into a ProtoConfig member
instead of a generic output parameter, so that the cookie code
is able to set everything in its own pcfg_ member.

reset_tls_crypt_server() has also been modified to simply use
the symmetric key if available (because the cookie code has
already extracted it).

The PsidCookie now take non-const buffers as input, because
with WKc-containing TLS crypt V2 packets we need to remove the
WKc after processing the packet.

Signed-off-by: Razvan Cojocaru <razvan.cojocaru@openvpn.com>