Check for libcap at the cmake setup step
We’re not currently checking if libcap is present on the system
before building, but core code requires the presence of
capabilities.h.

So the build will just fail halfway through if libcap is not
installed, and then the developer is left with figuring out
which package sys/capability.h belongs to, and then install it
and resume the build.

This patch adds a helper function that explicitly checks that
libcap is available on the system.

Jira: PG-262
Signed-off-by: Razvan Cojocaru <razvan.cojocaru@openvpn.com>