WFP: give access to local DNS only, when requested
Instead of not adding packet filter rules at all, when outside DNS is to
be blocked, but access to DNS on loopback is requested to be still allowed,
we add the packetfilter rules, but not the ones that block local udp/53.

This way we make sure that only access to local DNS is given, instead of
access to any DNS. This is more precisely what the user requested in the
first place, and also matches previous behavior with --dhcp-options,
where --block-outside-dns was always active implicitely.

The observation that the Umbrella DNS client also uses udp/53 to forward
queries to the recursive resolver could not be reproduced again. So, the
requirement to not block udp/53 at all is no longer given and thus the
rules for DNS on other interfaces is not an issue really.

Jira: OVPN3-1333
Signed-off-by: Heiko Hund <heiko@openvpn.net>