Clean up to_string.hpp
Make the openvpn::to_string() function more selective about what it will
be enabled for. This is to prevent the function from being used in
contexts where it is not appropriate, such as when the type being
converted to a string is already handled by the std::to_string()
function.

Now it will reuse std::to_string() for types that it can handle, and
only provide a fallback for types that std::to_string() does not
support.

Signed-off-by: Charlie Vigue <charlie.vigue@openvpn.com>