agent win: make event cancellation logic more robust
The "client destroy" event is created by the client, and its handle
is passed to the agent. The agent duplicates the handle into its own
process and listens for the event. When the client signals the event,
the callback is invoked in the agent, which then closes the tun.

In the /tun-setup request handler, the agent closes any existing event
handle before using the new one provided by the client. Closing the
handle invokes the callback with the error code `operation_aborted`.
In this case, we know the event was not signaled by the client and
therefore do not close the tun.

However, there are rare cases where the /tun-setup request and the
"client destroy" signal arrive simultaneously and are executed in
the wrong order. In this case, the agent creates the tun, returns
its handle to the client, and then closes the event handle. This
results in a fatal "handle is invalid" error on the client side.

The safeguard in the agent to detect this case fails because the
event was already signaled—so when the callback is invoked, it does
not return `operation_aborted`.

Fix this by introducing a lightweight wrapper around the event handle.
The wrapper also includes a `shared_ptr` to a `bool` flag. When a
handle is assigned, a new shared pointer is created (refcount = 1).
When setting up the callback, we capture the shared pointer in the
lambda (refcount = 2). When the handle is closed in the agent, we
set the flag to `true`. Later, when assigning a new handle, we create
a new shared pointer, but the old lambda still holds the original one
with the flag set to `true`.

When the lambda is invoked, it checks the flag. If it's `true`, we
know the operation was aborted (i.e., the callback is stale) and we
do not close the tun—thus avoiding the "handle is invalid" error
on the client side.

Jira: OVPN3-1337

Signed-off-by: Lev Stipakov <lev@openvpn.net>