Render vpn interface index as string in Connected event
Commit

  3b6ba2 ("win: provide VPN adapter index to Connected event")

has added a propagation of a VPN interface index to the Connected
event, but by some reasons in the merged commit the index is
missing from render() output. As it turned out, this is the only
way for apps to access it without refactoring event propagation
between core and app.

Add vpn interface index to render() output if it is not equal to -1.

Note that while this code executes on all platforms,
it solves Windows-specific issue - VPN adapter index is required
on Windows to exclude VPN interface when looking for the best host route.

Jira: OVPN3-1339

Signed-off-by: Lev Stipakov <lev@openvpn.net>