Cleanup control_net_recv overload.
This commit implements the overload in terms of the existing
control_net_recv function so the code is not duplicated. The one line
implementation should likely boil down to the same assembly without
the copy + paste code.

Also reordered the overload and added doxygen so the dependent overload
follows the one that it calls.

Signed-off-by: Charlie Vigue <charlie.vigue@openvpn.com>