Also print key agreement when printing negotiated details
With TLS 1.0 to 1.2, the used key agreement was depended on the certificates
themselves. With TLS 1.3 this is no longer the case but basically always
X25519 was used.  So this information has not been very interesting so far.

With OpenSSL 3.5.0 and the new X25519MLKEM768 hybrid key agreement, the used
key agreement group actually becomes interesting information.

This commit adds printing this information for OpenSSL 3.0.0+ and uses
a compat version for OpenSSL 3.0-3.1 to avoid an additional ifdef in the
code itself.

This also renames SSL Handshake to TLS Handshake since we never actually
supported SSL and renames c_ssl to ssl as it is no longer const.

Jira: OVPN3-1341
Signed-off-by: Arne Schwabe <arne@rfc2549.org>