Add a noexcept string formatter using vformat
This commit adds a new function `string::safe_format` to the `string`
namespace, which is a noexcept wrapper for `vformat`.

Also includes a helper function args_to_string that might be useful
in other contexts.

Unit tests included for both functions.

Now requires libfmt (libfmt-dev on Debian/Ubuntu) to be installed.

Signed-off-by: Charlie Vigue <charlie.vigue@openvpn.com>