dns: Refactor DnsServer::{dnssec,transport}_string() implementation
These two helper methods required an object to be instantiated, but at
the same time it needed the dnssec or transport value to be passed to
it.

This refacords these methods to be static methods which can be used
directly without first instantiating an object.  In addition, add an
overloaded method for each of them, not requiring an argument - thus
returning the dnssec/transport string represtentation of the
stored value in the object.

While touching these code paths, add some more doxygen documentation.

Signed-off-by: David Sommerseth <davids@openvpn.net>