linux/dco: switch to supporting the new 'ovpn' kernel module
The 'ovpn' kernel module has been merged upstream and it is on
it's way to Linux 6.16.

Meanwhile, we have already developed a 'ovpn-backports' package
to help people compile and install it on systems running an older
kernel.

This patch drops support for the older ovpn-dco-v2 and introduces
support for ovpn.

Compile defines remain the same, so nothing should be changed
on the client side, other than installing the correct module.

While at it, also update the Linux DCO related instruction in
the README.md file.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>