win: always set DNS addresses at the VPN adapter
Previous to this the addresses were only set to prevent a corner case in
conjunction with resolvers on loopback. However search domains will not
work without the adapter DNS servers set, openvpn 2 does set them
because of this. Learnings from a previous issue suggests that search
domains are being actively used in the field, and it would be a shame if
they don't work as expected with --dns.

Another thing this commit adds, is the flushing of the resolver cache
after --dns has been applied to the system. Previously this has only
happened for --dhcp-option.

Jira: OVPN3-1357
Signed-off-by: Heiko Hund <heiko@openvpn.net>