ovpn_dco_linux: drop unused enum clashing with kernel headers
enum ovpn_ifla_attrs and enum ovpn_mode are still clashing with
linux-6.16 headers (due to include ordering), despite a dirst
attempt to fix this issue.

However, these defines are not used at all in core because we
always create a DCO interface with the default params.

Fix the clash for good by deleting the offending enums.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>