Core: Add optional_ref class header and unit tests
Adds a new header file `optional_ref.hpp` that implements an
optional reference wrapper, similar to `std::optional`, but for
references. This class allows for the use of references in a way that
can be checked for existence without copying the underlying object.

This commit also includes unit tests to verify the functionality of
the class.

Signed-off-by: Charlie Vigue <charlie.vigue@openvpn.com>